Refactoring SumTest final version

+++ src/main/controller/MockSum.java
@@ -1,5 +1,5 @@
 package main.controller;
 
 public interface MockSum {
-    public float auxiliarSum(float x, float y); // Funció sense definir que modifiquem el comportament al test
+    public double auxiliarSum(double x, double y);
 }
